requirements.txt
torch
numpy
opencv-python
matplotlib
Pillow
tqdm
torchvision
scikit-learn
glob2
